Add files via upload
This file gives a brief introduction of the dataset related to national leaders across the world. The charts histogram of leadership duration, numbers of countries in each continent. It also plot some figure in terms of the start year, end year, duration, entry and exit mode of top 20 longest service leader. Most Importantly, it visualizes leadership transition in each country.